Move codon conversion functions into convert.codon module
